# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Vaishnavi Residency Hotel Management API
 * OpenAPI spec version: 0.1.0
 */
import type { RoomStatus } from './roomStatus';

export interface Room {
  id: number;
  roomNumber: string;
  type: string;
  rentPerDay: number;
  status: RoomStatus;
  /** @nullable */
  floor?: string | null;
  /** @nullable */
  description?: string | null;
  createdAt?: string;
}